requirements.txt
anthropic>=0.40.0
supabase>=2.9.0
feedparser>=6.0.11
httpx>=0.27.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.1
jinja2>=3.1.4
pydantic>=2.9.0
streamlit>=1.40.0
pandas>=2.2.0
